+++ src/clua-grpc-machine.cpp
@@ -40,8 +40,10 @@ void cm_delete(cm_grpc_machine_stub *ptr) {
 static int grpc_machine_class_get_default_config(lua_State *L) {
     auto &managed_grpc_stub =
         clua_check<clua_managed_cm_ptr<cm_grpc_machine_stub>>(L, lua_upvalueindex(1), lua_upvalueindex(2));
-    auto &managed_default_config = clua_push_to(L, clua_managed_cm_ptr<const cm_machine_config>(nullptr));
-    TRY_EXECUTE(cm_grpc_get_default_config(managed_grpc_stub.get(), &managed_default_config.get(), err_msg));
+    auto &managed_default_config =
+        clua_push_to(L, clua_managed_cm_ptr<const cm_machine_config>(nullptr), lua_upvalueindex(2));
+    TRY_EXECUTE_CTXIDX(cm_grpc_get_default_config(managed_grpc_stub.get(), &managed_default_config.get(), err_msg),
+        lua_upvalueindex(2));
     clua_push_cm_machine_config(L, managed_default_config.get());
     managed_default_config.reset();
     return 1;
@@ -53,8 +55,11 @@ static int grpc_machine_class_verify_access_log(lua_State *L) {
     lua_settop(L, 2);
     auto &managed_grpc_stub =
         clua_check<clua_managed_cm_ptr<cm_grpc_machine_stub>>(L, lua_upvalueindex(1), lua_upvalueindex(2));
-    auto &managed_log = clua_push_to(L, clua_managed_cm_ptr<cm_access_log>(clua_check_cm_access_log(L, 1)));
-    TRY_EXECUTE(cm_grpc_verify_access_log(managed_grpc_stub.get(), managed_log.get(), lua_toboolean(L, 2), err_msg));
+    auto &managed_log =
+        clua_push_to(L, clua_managed_cm_ptr<cm_access_log>(clua_check_cm_access_log(L, 1)), lua_upvalueindex(2));
+    TRY_EXECUTE_CTXIDX(
+        cm_grpc_verify_access_log(managed_grpc_stub.get(), managed_log.get(), lua_toboolean(L, 2), err_msg),
+        lua_upvalueindex(2));
     managed_log.reset();
     lua_pop(L, 1); // managed pointer
     lua_pushnumber(L, 1);
@@ -73,8 +78,9 @@ static int grpc_machine_class_verify_state_transition(lua_State *L) {
     cm_hash target_hash{};
     clua_check_cm_hash(L, 3, &target_hash);
     const bool one_based = lua_toboolean(L, 4);
-    TRY_EXECUTE(cm_grpc_verify_state_transition(managed_grpc_stub.get(), &root_hash, managed_log.get(), &target_hash,
-        one_based, err_msg));
+    TRY_EXECUTE_CTXIDX(cm_grpc_verify_state_transition(managed_grpc_stub.get(), &root_hash, managed_log.get(),
+                           &target_hash, one_based, err_msg),
+        lua_upvalueindex(2));
     managed_log.reset();
     lua_pop(L, 1);        // managed pointer
     lua_pushnumber(L, 1); // result
@@ -86,7 +92,8 @@ static int grpc_machine_class_get_x_address(lua_State *L) {
     auto &managed_grpc_stub =
         clua_check<clua_managed_cm_ptr<cm_grpc_machine_stub>>(L, lua_upvalueindex(1), lua_upvalueindex(2));
     uint64_t reg_address{};
-    TRY_EXECUTE(cm_grpc_get_x_address(managed_grpc_stub.get(), luaL_checkinteger(L, 1), &reg_address, err_msg));
+    TRY_EXECUTE_CTXIDX(cm_grpc_get_x_address(managed_grpc_stub.get(), luaL_checkinteger(L, 1), &reg_address, err_msg),
+        lua_upvalueindex(2));
     lua_pushinteger(L, static_cast<lua_Integer>(reg_address));
     return 1;
 }
@@ -97,7 +104,8 @@ static int grpc_machine_class_get_csr_address(lua_State *L) {
         clua_check<clua_managed_cm_ptr<cm_grpc_machine_stub>>(L, lua_upvalueindex(1), lua_upvalueindex(2));
     uint64_t csr_address{};
     const CM_PROC_CSR csr = clua_check_cm_proc_csr(L, 1);
-    TRY_EXECUTE(cm_grpc_get_csr_address(managed_grpc_stub.get(), csr, &csr_address, err_msg));
+    TRY_EXECUTE_CTXIDX(cm_grpc_get_csr_address(managed_grpc_stub.get(), csr, &csr_address, err_msg),
+        lua_upvalueindex(2));
     lua_pushinteger(L, static_cast<lua_Integer>(csr_address));
     return 1;
 }
@@ -107,7 +115,8 @@ static int grpc_machine_class_get_dhd_h_address(lua_State *L) {
     auto &managed_grpc_stub =
         clua_check<clua_managed_cm_ptr<cm_grpc_machine_stub>>(L, lua_upvalueindex(1), lua_upvalueindex(2));
     uint64_t dhd_h_address{};
-    TRY_EXECUTE(cm_grpc_dhd_h_address(managed_grpc_stub.get(), luaL_checkinteger(L, 1), &dhd_h_address, err_msg));
+    TRY_EXECUTE_CTXIDX(cm_grpc_dhd_h_address(managed_grpc_stub.get(), luaL_checkinteger(L, 1), &dhd_h_address, err_msg),
+        lua_upvalueindex(2));
     lua_pushinteger(L, static_cast<lua_Integer>(dhd_h_address));
     return 1;
 }