+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix Lua bind of static methods for remote machines
+- Fix rollup-memory-range decoding of empty payloads at end of file
+
 ## [0.10.0] - 2022-07-04
 ### Added